backend/services/ai-service/requirements.txt
# ai-service 依赖
# 作为 telegram-service 子模块，复用其 telegram 依赖

# 数据库
psycopg[binary]>=3.1.0

# 环境变量
python-dotenv>=1.0.0

# Gemini SDK
google-generativeai>=0.8.0
